Add a run() method for a simple chat loop
